Add validations

+++ actors-backend/routers/actors.js
@@ -1,4 +1,4 @@
-import {verifyToken} from "../utils/authentication.js";
+import {verifyAdmin, verifyToken} from "../utils/authentication.js";
 import {createActor, deleteActor, getActor, getActors, modifyActor} from "../controllers/actors.js";
 
 export const actorsRouter = async (req, res) => {
@@ -19,16 +19,19 @@ export const actorsRouter = async (req, res) => {
         req.params.id = id;
         console.log(`DELETE /actors/${id}`);
         verifyToken(req);
+        verifyAdmin(req);
         await deleteActor(req, res);
     } else if (req.fullUrl.pathname === '/' && req.method === 'POST') {
         console.log("POST /actors");
         verifyToken(req);
+        verifyAdmin(req);
         await createActor(req, res);
     } else if (regexRouteWithId.test(req.fullUrl.pathname) && req.method === 'PUT') {
         let id = parseInt(req.fullUrl.pathname.match(regexRouteWithId)[1], 10);
         req.params.id = id;
         console.log(`PUT /actors/${id}`);
         verifyToken(req);
+        verifyAdmin(req);
         await modifyActor(req, res);
     } else {
         req.handled = false;
